backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-genai==0.3.0
python-dotenv==1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn==21.2.0
